+++ desktop/apps/auction/client/analytics_middleware.js
